+++ src/cspEngine/models/Environment.java
@@ -1,4 +1,4 @@
-package models;
+package cspEngine.models;
 
 import java.util.ArrayList;
 import java.util.List;
