+++ src/graphEngine/Reference.java
@@ -1,89 +1,9 @@
-package automaton;
+package graphEngine;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Scanner;
-
-class Node {
-	private String name;
-	private List<Path> paths;
-	private int distanceFromFinal;
-
-	public Node(String name, int distanceFromFinal) {
-		this.name = name;
-		this.distanceFromFinal = distanceFromFinal;
-		paths = new ArrayList<>();
-	}
-
-	public void point(String edges) {
-		String[] properties = edges.split(" ");
-		for (int i = 0; i < properties.length - 1; i = i + 2)
-			paths.add(new Path(properties[i], properties[i + 1]));
-	}
-
-	public Node(Node node) {
-		name = node.name();
-		paths = new ArrayList<>(node.paths());
-		distanceFromFinal = node.dist();
-	}
-
-	public String name() {
-		return name;
-	}
-
-	public int dist() {
-		return distanceFromFinal;
-	}
-
-	public List<Path> paths() {
-		return paths;
-	}
-
-	public String follow(String path) {
-		for (Path p : paths)
-			if (p.path().equalsIgnoreCase(path))
-				return p.dest();
-		return null;
-	}
-
-	@Override
-	public String toString() {
-		String str = "";
-		str += name + ":\n";
-		for (Path p : paths)
-			str += "\t" + p.toString() + "\n";
-		return str;
-	}
-
-	@Override
-	public Node clone() {
-		return new Node(this);
-	}
-}
-
-class Path {
-	private String path;
-	private String dest;
-
-	public Path(String path, String dest) {
-		this.path = path;
-		this.dest = dest;
-	}
-
-	public String path() {
-		return path;
-	}
-
-	public String dest() {
-		return dest;
-	}
-
-	@Override
-	public String toString() {
-		return "" + path +" "+ dest;
-	}
-}
 
 public class Reference {
 	public static HashMap<String, Node> map;
@@ -729,193 +649,3 @@ public class Reference {
 	}
 }
 
-class Driver {
-	public static void main(String[] args) {
-		Reference.init();
-		System.out.println("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n");
-		while (startMenu() == 0)
-			mainGame();
-
-	}
-
-	public static boolean mainGame() {
-		Scanner input = new Scanner(System.in);
-		String currentState = Reference.START_STATE;
-		boolean stepAllFlag = false;
-		int turn = 1;
-
-		while (!currentState.equals(Reference.END_STATE)) {
-			System.out.println("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n");
-			System.out.println("------------------");
-			System.out.println("\tTurn [" +turn+ "]\t");
-			System.out.println("------------------");
-			showState(currentState);
-			System.out.println("Input hint: <source> <destination> <entity1> <entity2> ... <entityN>");
-			System.out.print("==> ");
-			String in = "";
-
-			if (!stepAllFlag) {
-				in = input.nextLine();
-				if (!in.equalsIgnoreCase("stepall") && !in.equalsIgnoreCase("step"))
-					in = preprocessInput(in);
-			}
-
-			if (in.equalsIgnoreCase("stepall"))
-				stepAllFlag = true;
-			if (in.equalsIgnoreCase("step") || stepAllFlag) {
-				in = Reference.smartChoice(currentState);
-				System.out.println(in);
-			}
-
-			in = Reference.get(currentState).follow(in);
-
-			if (in == null) {
-				System.out.println("You can't do that.\n");
-				continue;
-			}
-			currentState = in;
-			turn++;
-		}
-
-		System.out.println("You solved the game!");
-		return true;
-	}
-
-	public static void showState(String currentState) {
-		String envs = Reference.corres.get(currentState);
-		String[] envs_ = envs.split("\\|");
-		for (int i = 0; i < envs_.length; i++) {
-			if (i == 0) {
-				System.out.println("Earth:");
-				for (int j = 0; j < envs_[i].length(); j++)
-					if (!(envs_[i].charAt(j)+"").equals(" "))
-						System.out.println("\t" + toCleanCharacter("" + envs_[i].charAt(j)));
-				System.out.println();
-			}
-			else if (i == 1) {
-				System.out.println("Spaceship:");
-				for (int j = 0; j < envs_[i].length(); j++)
-					if (!(envs_[i].charAt(j)+"").equals(" "))
-						System.out.println("\t" + toCleanCharacter("" + envs_[i].charAt(j)));
-				System.out.println();
-			}
-			else {
-				System.out.println("Mars:");
-				for (int j = 0; j < envs_[i].length(); j++)
-					if (!(envs_[i].charAt(j)+"").equals(" "))
-						System.out.println("\t" + toCleanCharacter("" + envs_[i].charAt(j)));
-				System.out.println();
-			}
-		}
-	}
-
-	public static String toCharacter(String ch) {
-		if (ch.equalsIgnoreCase("s") || ch.equalsIgnoreCase("scientist"))
-			return "S";
-		if (ch.equalsIgnoreCase("h") || ch.equalsIgnoreCase("human"))
-			return "H";
-		if (ch.equalsIgnoreCase("l") || ch.equalsIgnoreCase("lion"))
-			return "L";
-		if (ch.equalsIgnoreCase("c") || ch.equalsIgnoreCase("cow"))
-			return "C";
-		if (ch.equalsIgnoreCase("g") || ch.equalsIgnoreCase("grain"))
-			return "G";
-		return "";
-	}
-
-	public static String toCleanCharacter(String ch) {
-		if (ch.equalsIgnoreCase("s") || ch.equalsIgnoreCase("scientist"))
-			return "Scientist";
-		if (ch.equalsIgnoreCase("h") || ch.equalsIgnoreCase("human"))
-			return "Human";
-		if (ch.equalsIgnoreCase("l") || ch.equalsIgnoreCase("lion"))
-			return "Lion";
-		if (ch.equalsIgnoreCase("c") || ch.equalsIgnoreCase("cow"))
-			return "Cow";
-		if (ch.equalsIgnoreCase("g") || ch.equalsIgnoreCase("grain"))
-			return "Grain";
-		return "";
-	}
-
-	public static String toEnv(String name) {
-		if (name.equalsIgnoreCase("earth") || name.equalsIgnoreCase("e"))
-			return "E";
-		if (name.equalsIgnoreCase("spaceship") || name.equalsIgnoreCase("s"))
-			return "S";
-		if (name.equalsIgnoreCase("mars") || name.equalsIgnoreCase("m"))
-			return "M";
-		return null;
-	}
-
-	public static String preprocessInput(String input) {
-		String[] in_ = input.split(" ");
-
-		if ((""+in_[0].charAt(0)).equals("<") && (""+in_[0].charAt(in_[0].length()-1)).equals(">"))
-			return in_[0];
-
-		input = "<" + toEnv(in_[0]) + toEnv(in_[1]);
-		input += "-";
-
-		for (int i = 2; i < in_.length; i++) {
-			String s = in_[i];
-			input += toCharacter(s);
-		}
-		input += ">";
-		return input;
-	}
-
-	public static int startMenu() {
-		Scanner in = new Scanner(System.in);
-		int choice = 1;
-
-		while (choice == 1) {
-			System.out.println("===============================================");
-			System.out.println(" T H E  T R A V E L L I N G  S C I E N T I S T ");
-			System.out.println("===============================================");
-
-			System.out.println("Main Menu :");
-			System.out.println("[0]\t Play");
-			System.out.println("[1]\t How to play");
-			System.out.println("[2]\t Exit\n\n");
-
-			System.out.print(">> ");
-			choice = in.nextInt();
-
-			if (choice == 0) {
-				return 0;
-			}
-			if (choice == 1) {
-				help();
-			}
-			if (choice == 2) {
-				return 2;
-			}
-		}
-		return -1;
-	}
-
-	public static void help() {
-		System.out.println("\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n");
-		System.out.println("=====================");
-		System.out.println(" H O W  T O  P L A Y ");
-		System.out.println("=====================");
-
-		String help = "\n";
-		help += "1\t HOW TO PLAY\n";
-		help += "\tEvery turn, you are asked where to get people and " +
-				"where \n\tto drop them off, and who are involved in the migration.\n\n" +
-				"\tInput format : \"<source> <destination> <entity1> <entity2> ... <entityN>\"";
-		help += "\n\n\n";
-		help += "2\t BACKSTORY\n";
-		help += " The Earth is slowly dying and a scientist wants to help build a new life on Mars.\n " +
-				"However, he only has a spaceship that was so tiny it could only carry him and two\n " +
-				"other items. He wants to transport five things that he thought were essential to\n " +
-				"start a new life on Mars: two humans, one lion, one cow and one bag of grain.\n " +
-				"Whenever the scientist is not around, either human would kill the lion out of fear\n " +
-				"or eat the cow out of hunger,the lion will eat the cow, and the cow will eat the\n " +
-				"grain. Only the scientist knows how to fly the spaceship. How will the scientist\n " +
-				"transport all the five items to Mars?\n\n\n\n\n\n\n\n\n\n";
-
-		System.out.println(help);
-	}
-}
